requirements.txt
alembic==1.7.7
aniso8601==9.0.1
asttokens==2.0.5
asyncpg==0.25.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
decorator==5.1.1
executing==0.8.3
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Email==1.4.4
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
install==1.3.5
ipdb==0.13.9
ipython==8.1.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
protobuf==3.20.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.32
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
